I had the whole design backwards, IR shouldn't know about the different frontends (much like it doesn't know about the backends)
